+++ src/main/java/io/wispforest/owo/mixin/serialization/CachedRegistryInfoGetterAccessor.java
@@ -1,4 +1,4 @@
-package io.wispforest.owo.mixin;
+package io.wispforest.owo.mixin.serialization;
 
 import net.minecraft.registry.RegistryOps;
 import net.minecraft.registry.RegistryWrapper;
