+++ src/main/java/io/wispforest/owo/mixin/serialization/ComponentTypeBuilderMixin.java
@@ -1,9 +1,9 @@
-package io.wispforest.owo.mixin;
+package io.wispforest.owo.mixin.serialization;
 
 import io.wispforest.owo.serialization.OwoComponentTypeBuilder;
 import net.minecraft.component.ComponentType;
 import org.spongepowered.asm.mixin.Mixin;
 
 @Mixin(ComponentType.Builder.class)
-public class ComponentTypeBuilderMixin<T> implements OwoComponentTypeBuilder<T> {
+public abstract class ComponentTypeBuilderMixin<T> implements OwoComponentTypeBuilder<T> {
 }
