+++ src/main/java/io/wispforest/owo/mixin/serialization/NbtCompoundMixin.java
@@ -1,4 +1,4 @@
-package io.wispforest.owo.mixin;
+package io.wispforest.owo.mixin.serialization;
 
 import io.wispforest.endec.SerializationAttributes;
 import io.wispforest.endec.SerializationContext;
