+++ src/main/java/io/wispforest/owo/mixin/serialization/PacketByteBufMixin.java
@@ -1,4 +1,4 @@
-package io.wispforest.owo.mixin;
+package io.wispforest.owo.mixin.serialization;
 
 import io.wispforest.endec.Endec;
 import io.wispforest.endec.SerializationContext;
@@ -10,7 +10,7 @@ import org.spongepowered.asm.mixin.Mixin;
 
 @SuppressWarnings({"DataFlowIssue", "AddedMixinMembersNamePattern"})
 @Mixin(PacketByteBuf.class)
-public class PacketByteBufMixin implements EndecBuffer {
+public abstract class PacketByteBufMixin implements EndecBuffer {
     @Override
     public <T> void write(SerializationContext ctx, Endec<T> endec, T value) {
         endec.encodeFully(ctx, () -> ByteBufSerializer.of((PacketByteBuf) (Object) this), value);
