+++ src/main/java/io/wispforest/owo/serialization/CodecUtils.java
@@ -17,8 +17,8 @@ import io.wispforest.endec.format.forwarding.ForwardingSerializer;
 import io.wispforest.endec.format.gson.GsonDeserializer;
 import io.wispforest.endec.format.gson.GsonEndec;
 import io.wispforest.endec.format.gson.GsonSerializer;
-import io.wispforest.owo.mixin.ForwardingDynamicOpsAccessor;
-import io.wispforest.owo.mixin.RegistryOpsAccessor;
+import io.wispforest.owo.mixin.serialization.ForwardingDynamicOpsAccessor;
+import io.wispforest.owo.mixin.serialization.RegistryOpsAccessor;
 import io.wispforest.owo.serialization.endec.EitherEndec;
 import io.wispforest.owo.serialization.endec.MinecraftEndecs;
 import io.wispforest.owo.serialization.endec.StructEitherEndec;
