+++ src/main/resources/owo.mixins.json
@@ -4,16 +4,16 @@
   "package": "io.wispforest.owo.mixin",
   "compatibilityLevel": "JAVA_16",
   "mixins": [
-    "CachedRegistryInfoGetterAccessor",
     "ClientConnectionMixin",
-    "ComponentTypeBuilderMixin",
     "Copenhagen",
-    "DataResultMixin",
-    "DataResultMixin$DataResultErrorMixin",
-    "ForwardingDynamicOpsAccessor",
-    "NbtCompoundMixin",
-    "PacketByteBufMixin",
-    "RegistryOpsAccessor",
+    "serialization.CachedRegistryInfoGetterAccessor",
+    "serialization.ComponentTypeBuilderMixin",
+    "serialization.DataResultMixin",
+    "serialization.DataResultMixin$DataResultErrorMixin",
+    "serialization.ForwardingDynamicOpsAccessor",
+    "serialization.NbtCompoundMixin",
+    "serialization.PacketByteBufMixin",
+    "serialization.RegistryOpsAccessor",
     "serialization.ReadViewMixin",
     "serialization.WriteViewMixin",
     "ScreenHandlerInvoker",
